requirements.txt
torch==1.10
torchvision==0.10
opencv-python==4.10.0
pyqt5==5.15.9
qt-material==2.14
matplotlib==3.6.0
numpy==1.23.2
pillow==10.4.0
tqdm==4.67.1